requirements.txt
psycopg2-binary==2.8.6
pandas==1.2.3
matplotlib==3.3.4
